--- a/src/operations/xor.function.ts
+++ b/src/operations/xor.function.ts
@@ -11,16 +11,53 @@ export function xor<T>(...sets: ReadonlySet<T>[]): ReadonlySet<T>;
  * @description A ∆ B ≔ { x : (x ∈ A) ⊕ (x ∈ B) }
  */
 export function xor<T, S extends ReadonlySet<T>>(...sets: S[]): S {
-	const resultSet = new Set<T>(sets.shift());
-	const reusedElements = new Set<T>();
+	if (sets.length < 2) {
+		return new Set<T>(sets.shift()) as ReadonlySet<T> as S;
+	}
 
-	for (const set of sets) {
-		for (const element of set) {
-			if (resultSet.has(element)) {
-				resultSet.delete(element);
-				reusedElements.add(element);
-			} else if (!reusedElements.has(element)) {
-				resultSet.add(element);
+	const resultSet = new Set<T>();
+	const duplicates = new Set<T>();
+	const primarySet = sets.shift()!;
+	const secondarySet = sets.shift()!;
+
+	primaryLoop(primarySet, secondarySet);
+	primaryLoop(secondarySet, primarySet);
+	function primaryLoop(primary: S, secondary: S): void {
+		if (sets.length !== 0) {
+			for (const element of primary) {
+				if (!secondary.has(element)) {
+					resultSet.add(element);
+				} else {
+					duplicates.add(element);
+				}
+			}
+		} else {
+			for (const element of primary) {
+				if (!secondary.has(element)) {
+					resultSet.add(element);
+				}
+			}
+		}
+	}
+
+	sets.forEach(tertiaryLoop);
+	function tertiaryLoop(set: S, index: number): void {
+		if (index < sets.length - 1) {
+			for (const element of set) {
+				if (resultSet.has(element)) {
+					resultSet.delete(element);
+					duplicates.add(element);
+				} else if (!duplicates.has(element)) {
+					resultSet.add(element);
+				}
+			}
+		} else {
+			for (const element of set) {
+				if (resultSet.has(element)) {
+					resultSet.delete(element);
+				} else if (!duplicates.has(element)) {
+					resultSet.add(element);
+				}
 			}
 		}
 	}